requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.6
djangorestframework==3.15.2
psycopg==3.2.5
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
